third_party/blink/web_tests/platform/generic/http/tests/history/redirect-js-document-location-before-load-expected.txt

This page is the target of a redirect.

PASS: History item count should be 1 and is.


============== Back Forward List ==============
curr->  http://127.0.0.1:8000/history/resources/redirect-target.html#1
===============================================
